+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Bundle loaded with javascript for loading progress.
 - Bundle can be created after the server start and bundling info page is rendered (instead of webserver error).
 - Option to auto-generate component wrappers in the layout.
+- Custom files can be saved in RS directory.
 ### Breaking changes
 - Removed `Button` component.
 - Removed `res.render` function.